+++ treasury/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "treasury"
+version = "0.1.0"
+authors = ["Vicki Niu <vicki.niu@gmail.com>"]
+edition = "2018"
+
+[dependencies]
+failure = "0.1"
+byteorder = "1"
+merlin = "1.0.1"
+rand = "0.6"
+subtle = "2"
+curve25519-dalek = { version = "1.0.1", features = ["serde"] }
+
+[dependencies.bulletproofs]
+git = "https://github.com/dalek-cryptography/bulletproofs"
+branch = "develop"
+#path = "../../dalek/bulletproofs"
+features = ["yoloproofs"]
+
+[dependencies.spacesuit]
+path = "../spacesuit"
+
+[dependencies.zkvm]
+path = "../zkvm"
