+++ zkvm/src/lib.rs
@@ -12,7 +12,6 @@ mod prover;
 mod scalar_witness;
 mod signature;
 mod transcript;
-mod treasury;
 mod txlog;
 mod types;
 mod verifier;
@@ -26,7 +25,7 @@ pub use self::predicate::Predicate;
 pub use self::prover::Prover;
 pub use self::signature::{Signature, VerificationKey};
 pub use self::transcript::TranscriptProtocol;
-pub use self::txlog::{Entry, TxID, UTXO};
+pub use self::txlog::{Entry, TxID, TxLog, UTXO};
 pub use self::types::{Data, Item, Value, WideValue};
 pub use self::verifier::Verifier;
 pub use self::vm::{Tx, TxHeader, VerifiedTx};
